Add files via upload
This is the json file for the wokwi diagram which works with the sketch.ino file which shows how the project works virtually without the hardware. You can use the wokwi web and choose the arduino UNO board and paste the codes in their respective places and it should work.